+++ src/include/sof/lib_manager.h
@@ -155,9 +155,9 @@ struct processing_module;
  * Function is responsible to allocate module in available free memory and assigning proper address.
  * (WIP) These feature will contain module validation and proper memory management.
  */
-uint32_t lib_manager_allocate_module(struct processing_module *proc,
-				     struct comp_ipc_config *ipc_config,
-				     const void *ipc_specific_config, const void **buildinfo);
+uintptr_t lib_manager_allocate_module(struct processing_module *proc,
+				      struct comp_ipc_config *ipc_config,
+				      const void *ipc_specific_config, const void **buildinfo);
 
 /*
  * \brief Free module
