+++ src/include/sof/llext_manager.h
@@ -11,9 +11,9 @@
 struct comp_driver;
 struct comp_ipc_config;
 
-uint32_t llext_manager_allocate_module(struct processing_module *proc,
-				       struct comp_ipc_config *ipc_config,
-				       const void *ipc_specific_config, const void **buildinfo);
+uintptr_t llext_manager_allocate_module(struct processing_module *proc,
+					struct comp_ipc_config *ipc_config,
+					const void *ipc_specific_config, const void **buildinfo);
 int llext_manager_free_module(struct processing_module *proc,
 			      struct comp_ipc_config *ipc_config);
